refactor(wallet): Wallet.java 패키지 이동 vo -> domain
