+++ src/main/java/view/InputView.java
@@ -4,11 +4,11 @@ import static java.util.stream.Collectors.toList;
 
 import domain.Lotto;
 import domain.Lottos;
+import domain.Wallet;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
 import java.util.stream.Collectors;
-import vo.Wallet;
 
 public class InputView {
     private static final Scanner SCANNER = new Scanner(System.in);
