+++ src/test/java/domain/WalletTest.java
@@ -1,4 +1,4 @@
-package vo;
+package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
